State: Move profileLinks data-layer to appropariate directories
